% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnectdbs.R
\name{disconnectdbs}
\alias{disconnectdbs}
\title{Disconnect from the databases}
\usage{
disconnectdbs(inUrl, outUrl)
}
\arguments{
\item{inUrl}{JDBC connection URL for input data, defaults to the value of environment parameter IN_JDBC_URL}

\item{outUrl}{JDBC connection URL for output data, defaults to the value of environment parameter OUT_JDBC_URL}
}
\description{
Environment variables:
- Execution context:
     IN_JDBC_URL : JDBC connection URL for input data
     OUT_JDBC_URL : JDBC connection URL for output results
}

