% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchData.R
\name{fetchData}
\alias{fetchData}
\title{Fetch data from the input database}
\usage{
fetchData(query, inFormat, conn)
}
\arguments{
\item{query}{The SQL query to execute on the input database, defaults to the value of environment parameter PARAM_query}

\item{inFormat}{Hint for the exact shape of the data stored in the database. Possible values are PARTIAL_RESULTS, OTHER. Defaults to the value of environment parameter IN_FORMAT}

\item{conn}{The connection to the database, default to global variable in_conn}
}
\description{
Environment variables:
}
\details{
- Input Parameters:
     PARAM_query  : SQL query producing the dataframe to analyse
- Execution context:
     IN_FORMAT : Hint for the exact shape of the data stored in the database.
       Possible values are PARTIAL_RESULTS, OTHER
     IN_JDBC_DRIVER : class name of the JDBC driver for input data
     IN_JDBC_JAR_PATH : path to the JDBC driver jar for input data
     IN_JDBC_URL : JDBC connection URL for input data
     IN_JDBC_USER : User for the database connection for input data
     IN_JDBC_PASSWORD : Password for the database connection for input data
     IN_JDBC_SCHEMA : Optional schema by default for the database connection for input data
}

