% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveError.R
\name{saveError}
\alias{saveError}
\title{Save an error into the output database}
\usage{
saveError(error, outputFile, errorFile, jobId, node, resultTable, fn, conn)
}
\arguments{
\item{error}{The error to record.}

\item{outputFile}{File containing the output of R, default to the value of environment variable OUTPUT_FILE.}

\item{jobId}{ID of the job, defaults to the value of environment parameter JOB_ID}

\item{node}{Node used for the execution of the script, defaults to the value of environment parameter NODE}

\item{resultTable}{Name of the result table, defaults to the value of environment parameter RESULT_TABLE}

\item{fn}{Hint about the function used to produce the data}

\item{conn}{The connection to the database, default to global variable out_conn}

\item{error}{File containing the error of R, default to the value of environment variable ERROR_FILE.}
}
\description{
Environment variables:
}
\details{
- Execution context:
     JOB_ID : ID of the job
     NODE : Node used for the execution of the script
     RESULT_TABLE: Name of the result table, defaults to 'job_result'
     OUT_JDBC_DRIVER : class name of the JDBC driver for output results
     OUT_JDBC_JAR_PATH : path to the JDBC driver jar for output results
     OUT_JDBC_URL : JDBC connection URL for output results
     OUT_JDBC_USER : User for the database connection for output results
     OUT_JDBC_PASSWORD : Password for the database connection for output results
     OUTPUT_FILE : File containing the output of R
     ERROR_FILE : File containing the errors of R
}

