% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveResults.R
\name{saveResults}
\alias{saveResults}
\title{Save the results into the output database}
\usage{
saveResults(results)
}
\arguments{
\item{results}{The results to store in the database. The following types are supported: data frame, matrix, string.}
}
\description{
Environment variables:
}
\details{
- Execution context:
     JOB_ID : ID of the job
     NODE : Node used for the execution of the script
     RESULT_TABLE: name of the result table, defaults to 'job_result'
     OUT_JDBC_DRIVER : class name of the JDBC driver for output results
     OUT_JDBC_JAR_PATH : path to the JDBC driver jar for output results
     OUT_JDBC_URL : JDBC connection URL for output results
     OUT_JDBC_USER : User for the database connection for output results
     OUT_JDBC_PASSWORD : Password for the database connection for output results
}

