% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect2indb.R
\name{connect2indb}
\alias{connect2indb}
\title{Connect to the input database}
\usage{
connect2indb()
}
\description{
Environment variables:
}
\details{
- Execution context:
     IN_JDBC_DRIVER : class name of the JDBC driver for input data
     IN_JDBC_JAR_PATH : path to the JDBC driver jar for input data
     IN_JDBC_URL : JDBC connection URL for input data
     IN_JDBC_USER : User for the database connection for input data
     IN_JDBC_PASSWORD : Password for the database connection for input data
}

