% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect2outdb.R
\name{connect2outdb}
\alias{connect2outdb}
\title{Connect to the output database}
\usage{
connect2outdb()
}
\description{
Environment variables:
}
\details{
- Execution context:
     OUT_JDBC_DRIVER : class name of the JDBC driver for output results
     OUT_JDBC_JAR_PATH : path to the JDBC driver jar for output results
     OUT_JDBC_URL : JDBC connection URL for output results
     OUT_JDBC_USER : User for the database connection for output results
     OUT_JDBC_PASSWORD : Password for the database connection for output results
}

