% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveResults.R
\name{saveResults}
\alias{saveResults}
\title{Save the results into the output database}
\usage{
saveResults(results, jobId, node, resultTable, outFormat, shape, fn, conn)
}
\arguments{
\item{results}{The results to store in the database. The following types are supported: data frame, matrix, string.}

\item{jobId}{ID of the job, defaults to the value of environment parameter JOB_ID}

\item{node}{Node used for the execution of the script, defaults to the value of environment parameter NODE}

\item{resultTable}{Name of the result table, defaults to the value of environment parameter RESULT_TABLE}

\item{outFormat}{Format requested for the output, default to value of environment parameter OUT_FORMAT}

\item{shape}{Hint about the shape of the data. The following shapes are supported: string, pfa_json, pfa_yaml,
r_dataframe_intermediate, r_dataframe_columns, r_matrix, r_other_intermediate}

\item{fn}{Hint about the function used to produce the data,  defaults to the value of environment parameter FUNCTION}

\item{conn}{The connection to the database, default to global variable out_conn}
}
\description{
Environment variables:
}
\details{
- Execution context:
     JOB_ID : ID of the job
     NODE : Node used for the execution of the script
     OUT_FORMAT : Hint for the exact shape of the Json stored in the database.
       Current values are PARTIAL_RESULTS, PRESENTATION
     RESULT_TABLE: Name of the result table, defaults to 'job_result'
     OUT_JDBC_DRIVER : class name of the JDBC driver for output results
     OUT_JDBC_JAR_PATH : path to the JDBC driver jar for output results
     OUT_JDBC_URL : JDBC connection URL for output results
     OUT_JDBC_USER : User for the database connection for output results
     OUT_JDBC_PASSWORD : Password for the database connection for output results
     FUNCTION: Name of the function executed
}

